+++ Extension/src/pages/filtering-log/components/RequestWizard/ruleCreators.ts
@@ -300,6 +300,26 @@ export const createExceptionRemoveHeaderRules = (event: FilteringLogEvent): stri
     return patterns;
 };
 
+/**
+ * Creates exception rules for urltransform modifier.
+ *
+ * @param event Filtering log event.
+ *
+ * @returns Array of exception rules.
+ */
+export const createExceptionUrlTransformRules = (event: FilteringLogEvent): string[] => {
+    const { frameDomain, requestRule } = event;
+    const patterns = [];
+
+    if (requestRule && requestRule.modifierValue) {
+        patterns.push(getUnblockDomainRule(frameDomain, `${NetworkRule.OPTIONS.URLTRANSFORM}=${requestRule.modifierValue}`));
+    }
+
+    patterns.push(getUnblockDomainRule(frameDomain, NetworkRule.OPTIONS.URLTRANSFORM));
+
+    return patterns;
+};
+
 export const createExceptionCspRules = (event: FilteringLogEvent): string[] => {
     const { frameDomain, requestRule } = event;
     const patterns = [];